test files of util moved to lib/util
